+++ app/src/main/java/com/example/schedule_manager/MainActivity.java
@@ -2,19 +2,11 @@ package com.example.schedule_manager;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.view.View;
 import android.widget.Button;
-import androidx.appcompat.app.AppCompatActivity;
 
-import android.database.Cursor;
-import android.database.SQLException;
-import android.database.sqlite.SQLiteDatabase;
-import android.os.Bundle;
-import android.view.View;
-import android.widget.Button;
-import android.widget.EditText;
-import android.widget.TextView;
-import java.io.IOException;
+import com.example.schedule_manager.adminUI.AdminLoginActivity;
+import com.example.schedule_manager.userUI.UserLoginActivity;
+
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
