+++ app/src/main/java/com/example/schedule_manager/adminUI/AdminAddEmployeeFragment.java
@@ -1,4 +1,4 @@
-package com.example.schedule_manager;
+package com.example.schedule_manager.adminUI;
 
 import android.os.Bundle;
 
@@ -8,17 +8,19 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ArrayAdapter;
-import android.widget.AutoCompleteTextView;
 import android.widget.Spinner;
 
-import java.text.DateFormatSymbols;
-import java.util.ArrayList;
-import java.util.HashSet;
-import java.util.Set;
+import com.example.schedule_manager.R;
 
 public class AdminAddEmployeeFragment extends Fragment {
 
     Spinner jidSpinner;
+    String[] examples = {
+        "Mhxanikos",
+                "Epivlepon",
+                "Grammateia",
+                "grmParagwgis",
+    };
     ArrayAdapter<String> adapterJid;
 
 
@@ -36,18 +38,7 @@ public class AdminAddEmployeeFragment extends Fragment {
 
 
     private void jidPopulateSpinner() {
-         int i =0 ;
-         ArrayList<Vardies> vardies = (ArrayList<Vardies>) MainActivity.getVardiesList();
-         HashSet<String> vardiesSet = new HashSet<>();
-         for(Vardies vrd: vardies){
-              vardiesSet.add(vrd.getEidikotita());
-         }
-         String[] examples2 = new String[vardiesSet.size()];
-         for(String eidik : vardiesSet){
-             examples2[i] = eidik;
-             i++;
-         }
-         adapterJid = new ArrayAdapter<>(getContext(), android.R.layout.simple_spinner_item,examples2);
+        adapterJid = new ArrayAdapter<>(getContext(), android.R.layout.simple_spinner_item,examples);
         adapterJid.setDropDownViewResource(android.R.layout.simple_spinner_dropdown_item);
         jidSpinner.setAdapter(adapterJid);
     }
