+++ app/src/main/java/com/example/schedule_manager/adminUI/AdminEmployeesFragment.java
@@ -1,4 +1,4 @@
-package com.example.schedule_manager;
+package com.example.schedule_manager.adminUI;
 
 import android.content.Intent;
 import android.os.Bundle;
@@ -11,7 +11,8 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.Button;
-import android.widget.ImageButton;
+
+import com.example.schedule_manager.R;
 
 
 public class AdminEmployeesFragment extends Fragment {
