+++ app/src/main/java/com/example/schedule_manager/adminUI/AdminNotificationsFragment.java
@@ -1,4 +1,4 @@
-package com.example.schedule_manager;
+package com.example.schedule_manager.adminUI;
 
 import android.os.Bundle;
 
@@ -7,6 +7,8 @@ import androidx.fragment.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+
+import com.example.schedule_manager.R;
 
 public class AdminNotificationsFragment extends Fragment {
     @Override
