+++ app/src/main/java/com/example/schedule_manager/adminUI/AdminProfileFragment.java
@@ -1,6 +1,5 @@
-package com.example.schedule_manager;
+package com.example.schedule_manager.adminUI;
 
-import android.content.Intent;
 import android.os.Bundle;
 
 import androidx.fragment.app.Fragment;
@@ -9,6 +8,8 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.TextView;
+
+import com.example.schedule_manager.R;
 
 public class AdminProfileFragment extends Fragment {
     protected TextView username, specialty;
