+++ app/src/main/java/com/example/schedule_manager/adminUI/AdminShowEmployeesListAdapter.java
@@ -1,4 +1,4 @@
-package com.example.schedule_manager;
+package com.example.schedule_manager.adminUI;
 
 import android.content.Context;
 import android.view.LayoutInflater;
@@ -9,6 +9,9 @@ import android.widget.TextView;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
+
+import com.example.schedule_manager.Ergazomenoi;
+import com.example.schedule_manager.R;
 
 import java.util.ArrayList;
