+++ app/src/main/res/layout/admin_edit_employees_activity.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".AdminEditEmployeesActivity">
+    tools:context=".adminUI.AdminEditEmployeesActivity">
 
     <FrameLayout
         android:id="@+id/edit_admin_frame"
