+++ app/src/main/res/layout/admin_edit_info_employee_fragment.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".AdminEditInfoEmployeeFragment">
+    tools:context=".adminUI.AdminEditInfoEmployeeFragment">
 
 
     <androidx.constraintlayout.widget.ConstraintLayout
