+++ app/src/main/res/layout/admin_login_activity.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".AdminLoginActivity">
+    tools:context=".adminUI.AdminLoginActivity">
 
     <LinearLayout
         android:layout_width="wrap_content"
