+++ app/src/main/res/layout/admin_remove_employee_fragment.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:layout_gravity="center"
     android:foregroundGravity="center"
-    tools:context=".AdminRemoveEmployeeFragment">
+    tools:context=".adminUI.AdminRemoveEmployeeFragment">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
