+++ app/src/main/res/layout/admin_shedule_saves_activity.xml
@@ -4,6 +4,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".AdminSheduleSaves">
+    tools:context=".adminUI.AdminSheduleSaves">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
