+++ app/src/main/res/layout/admin_shedule_show_activity.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".AdminSheduleShow">
+    tools:context=".adminUI.AdminSheduleShow">
 
     <LinearLayout
         android:id="@+id/linearLayout2"
