+++ app/src/main/res/layout/admin_show_employees_activity.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".AdminShowEmployeesActivity">
+    tools:context=".adminUI.AdminShowEmployeesActivity">
 
     <TextView
         android:id="@+id/textView3"
