+++ app/src/main/res/layout/admin_welcome_activity.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".AdminWelcomeActivity">
+    tools:context=".adminUI.AdminWelcomeActivity">
 
     <FrameLayout
         android:id="@+id/admin_frame"
