+++ app/src/main/res/layout/fragment_show_this_day_shedule.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ShowThisDaySheduleFragment">
+    tools:context=".adminUI.ShowThisDaySheduleFragment">
 
 
     <androidx.constraintlayout.widget.ConstraintLayout
