+++ app/src/main/res/layout/user_login_activity.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".UserLoginActivity">
+    tools:context=".userUI.UserLoginActivity">
 
     <LinearLayout
         android:layout_width="match_parent"
