+++ app/src/main/res/layout/user_profile_settings_activity.xml
@@ -4,6 +4,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".UserProfileSettings">
+    tools:context=".userUI.UserProfileSettings">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
