+++ app/src/main/res/layout/user_welcome_activity.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".UserWelcomeActivity">
+    tools:context=".userUI.UserWelcomeActivity">
 
     <FrameLayout
         android:id="@+id/main_frame"
